Ajuste para atualizar o saldo do caixa

+++ database/migrations/1559766979544_cashier_schema.js
@@ -15,7 +15,7 @@ class CashierSchema extends Schema {
         .onUpdate('CASCADE')
         .onDelete('CASCADE')
       table.string('description')
-      table.decimal('saldo').notNullable()
+      table.float('saldo').notNullable()
       table.timestamps()
     })
   }
